projects/application/backend-python/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic>=2.10
pydantic-settings>=2.7
python-keycloak>=4.0
httpx>=0.27
python-socketio[asyncio_client]>=5.11
anthropic>=0.40
openai>=1.50
python-jose[cryptography]>=3.3
python-multipart>=0.0.12
